+++ public/index.html
@@ -7,7 +7,6 @@
 		<link rel="stylesheet" type="text/css" href="src/css/index.css" />
 	</head>
 	<body>
-		<div id="app">
-		</div>
+		<div id="app"></div>
 	</body>
 </html>
