+++ src/App.tsx
@@ -1,10 +1,6 @@
-import { Calculator } from "./components/calculator";
+import { Calculator } from "components/calculator";
 function App() {
-  return (
-    <>
-      <Calculator />
-    </>
-  );
+  return <Calculator />;
 }
 
 export default App;
