+++ src/components/Operation.tsx
@@ -1,4 +1,4 @@
-import { OPERATORS } from "../constants";
+import { OPERATORS } from "constant";
 
 type Props = {
   operate: (operator: any) => void;
@@ -7,9 +7,9 @@ type Props = {
 export const Operation = ({ operate }: Props) => {
   return (
     <>
-      <div className="operations subgrid" onClick={operate}>
+      <div className="operations subgrid">
         {OPERATORS.map((operator) => (
-          <button key={operator} value={operator}>
+          <button key={operator} value={operator} onClick={operate}>
             {operator}
           </button>
         ))}
