+++ src/constant/index.ts
@@ -1,10 +1,6 @@
 export const OPERATORS = ['/', 'X', '-', '+', '='] as const;
 
-export const DIGITS = [9, 8, 7, 6, 5, 4, 3, 2, 1, 0] as const;
-
 export const DIGIT_MAX_LENGTH = 2;
-
-export const ZERO = '0';
 
 export const OPERATIONS = {
   ADD: '+',
