+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
-import './css/index.css';
-import reportWebVitals from './reportWebVitals';
+import App from 'App';
+import 'css/index.css';
+import reportWebVitals from 'reportWebVitals';
 
 const root = ReactDOM.createRoot(
   document.getElementById('app') as HTMLElement
